REFACTOR: move add list to package
