+++ src/drone_controller/util/list_operators.py
@@ -1,3 +1,13 @@
+def add_lists(first_list, second_list):
+    """
+    Adds two lists element wise
+    :param first_list:
+    :param second_list:
+    :return:
+    """
+    return [sum(x) for x in zip(first_list, second_list)]
+
+
 def subtract_lists(first_list, second_list):
     """
     Subtracts two lists
